refactor(message.ts): move JSON types out of message and into their own file
A file in src/Stratum/message.ts contained types for json, such as JSONArray, JSONValue, and
JSONObject. These have been moved out of /Stratum and into src/json.ts. Now src/Stratum/message.ts
contains types for Stratum messages, such as parameters and result. These are renamings of JSON
types and have been used throughout /Stratum instead of the json types.
