+++ src/Stratum/message.ts
@@ -1,15 +1,5 @@
-import * as Joi from 'joi'
+import * from '../json'
 
-export type JSONValue = null | string | number | boolean | JSONObject | JSONArray
+export type result = JSONValue
 
-export interface JSONObject {
-    [x: string]: JSONValue;
-}
-
-export interface JSONArray extends Array<JSONValue> { }
-
-let natural_number_schema = Joi.number().integer().min(0).required()
-
-export function is_natural_number(x): boolean {
-  return !natural_number_schema.validate(x).error
-}
+export type parameters = JSONArray
